refactor so no project header imports nostr.hpp
