+++ src/client/websocketpp_client.cpp
@@ -1,4 +1,6 @@
-#include "websocketpp_client.hpp"
+#include <mutex>
+
+#include "client/websocketpp_client.hpp"
 
 using namespace std;
